// app/src/shared/currency-code.enum.ts
export enum CurrencyCode {
    USDC = 'USDC',
    ETH = 'ETH',
    USDT = 'USDT',
    ARW = 'ARW',
    YAPYO = 'YAPYO',
    AVAX = 'AVAX',
    BCH = 'BCH',
    BNB = 'BNB',
    DAI = 'DAI',
    CGPT = 'CGPT',
    ALICE = 'ALICE',
    AITECH = 'AITECH',
    CAKE = 'CAKE',
    RIO = 'RIO',
    DEXE = 'DEXE',
    TOKEN = 'TOKEN',
    BRISE = 'BRISE',
    GAFI = 'GAFI',
    SFUND = 'SFUND',
    TLOS = 'TLOS',
    ORAI = 'ORAI',
    ERTHA = 'ERTHA',
    FORWARD = 'FORWARD',
    TLM = 'TLM',
    CREAM = 'CREAM',
    XCAD = 'XCAD',
    DCK = 'DCK',
    UNO = 'UNO',
    OFN = 'OFN',
    REEF = 'REEF',
    MTRG = 'MTRG',
    MARSH = 'MARSH',
    DPR = 'DPR',
    DMAIL = 'DMAIL',
    XTM = 'XTM',
    ECHO = 'ECHO',
    RVX = 'RVX',
    BTC = 'BTC',
    DASH = 'DASH',
    DOGE = 'DOGE',
    SHIB = 'SHIB',
    OM = 'OM',
    PEPE = 'PEPE',
    POL = 'POL',
    VERSE = 'VERSE',
    LINK = 'LINK',
    FET = 'FET',
    ONDO = 'ONDO',
    XCN = 'XCN',
    JASMY = 'JASMY',
    AAVE = 'AAVE',
    CRV = 'CRV',
    ENA = 'ENA',
    WLD = 'WLD',
    PAXG = 'PAXG',
    ETHFI = 'ETHFI',
    AIOZ = 'AIOZ',
    ACH = 'ACH',
    QNT = 'QNT',
    SUPER = 'SUPER',
    LDO = 'LDO',
    UNI = 'UNI',
    RSR = 'RSR',
    SKL = 'SKL',
    MASK = 'MASK',
    CRO = 'CRO',
    SAND = 'SAND',
    ENS = 'ENS',
    PENDLE = 'PENDLE',
    FTT = 'FTT',
    FLOKI = 'FLOKI',
    COTI = 'COTI',
    GRT = 'GRT',
    WAVES = 'WAVES',
    SAFE = 'SAFE',
    SWFTC = 'SWFTC',
    ANKR = 'ANKR',
    DYDX = 'DYDX',
    AXS = 'AXS',
    IMX = 'IMX',
    VENOM = 'VENOM',
    APE = 'APE',
    CHZ = 'CHZ',
    AMP = 'AMP',
    COMP = 'COMP',
    YFI = 'YFI',
    MNT = 'MNT',
    SUSHI = 'SUSHI',
    LRC = 'LRC',
    CVX = 'CVX',
    SAFFRONFI = 'SAFFRONFI',
    SOLAXY = 'SOLAXY',
    LTC = 'LTC',
    CRMS = 'CRMS',
    WIF = 'WIF',
    RAY = 'RAY',
    BONK = 'BONK',
    ORCA = 'ORCA',
    JUP = 'JUP',
    ZBCN = 'ZBCN',
    PYTH = 'PYTH',
    HNT = 'HNT',
    MEW = 'MEW',
    BOME = 'BOME',
    ZEUS = 'ZEUS',
    SLIM = 'SLIM',
    FCON = 'FCON',
    SAROS = 'SAROS',
    JTO = 'JTO',
    DRIFT = 'DRIFT',
    LIKE = 'LIKE',
    PRCL = 'PRCL',
    AAPLX = 'AAPLX',
    CRCLX = 'CRCLX',
    TON = 'TON',
    HMSTR = 'HMSTR',
    NOT = 'NOT',
    GAME = 'GAME',
    TRX = 'TRX',
    NFT = 'NFT',
    SUN = 'SUN',
    JST = 'JST',
    XMR = 'XMR',
}